# docs/swagger.yaml
definitions:
  internal_handlers.CreateRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  internal_handlers.TotalCostResponse:
    properties:
      total_cost:
        example: 498
        type: integer
    type: object
  internal_handlers.UpdateRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    required:
    - price
    - service_name
    - start_date
    type: object
  subscriptons-service_internal_models.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Список подписок пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subscriptons-service_internal_models.Subscription'
            type: array
      summary: Все подписки
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Новая подписка пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscriptons-service_internal_models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Полное удаление подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Детали подписки
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptons-service_internal_models.Subscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Частичное обновление подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptons-service_internal_models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Сумма подписок с фильтрацией по user_id, service_name, датам
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса (Yandex Plus)
        in: query
        name: service_name
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TotalCostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Общая стоимость подписок за период
      tags:
      - subscriptions
swagger: "2.0"